Restructure and always overwrite touchStart coords
