+++ src/board/utils.rs
@@ -1,0 +1,54 @@
+/// Determine the index of the empty field (`u8::MAX`) in a vector of fields.
+pub fn get_empty_field_idx(fields: &Vec<u8>) -> usize {
+    for (idx, &value) in fields.iter().enumerate() {
+        if value == u8::MAX {
+            return idx;
+        }
+    }
+
+    panic!("Could not find empty field!");
+}
+
+/// Get the left/top coordinates based on the index of a board field.
+pub fn get_left_top(idx: usize, width: usize, unit_size: usize) -> (usize, usize) {
+    let (row, col): (usize, usize) = get_row_col_from_idx(idx, width);
+    let left = col * unit_size;
+    let top = row * unit_size;
+
+    (left, top)
+}
+
+/// Get the row/column coordinates for a linear array representing a board.
+pub fn get_row_col_from_idx<T, U>(idx: T, width: T) -> (U, U)
+where
+    T: std::ops::Div<Output = T>,
+    T: std::ops::Rem<Output = T>,
+    T: Copy,
+    U: std::convert::From<T>,
+{
+    let row = idx / width;
+    let col = idx % width;
+
+    (row.into(), col.into())
+}
+
+/// Get the index into a linear array based on row/column coordinates.
+pub fn get_idx_from_row_col<T, U>(row: T, col: T, width: T) -> U
+where
+    T: std::ops::Mul<Output = T>,
+    T: std::ops::Add<Output = T>,
+    U: std::convert::From<T>,
+{
+    row.mul(width).add(col).into()
+}
+
+/// Check if row/column coordinates are within a field defined by width/height.
+pub fn in_bounds<T, U>(row: T, col: T, width: U, height: U) -> bool
+where
+    T: PartialOrd<T>,
+    T: PartialOrd<U>,
+    T: Default,
+{
+    let t_zero: T = T::default();
+    t_zero <= row && row < height && t_zero <= col && col < width
+}
