+++ src/slide_puzzle.rs
@@ -1,6 +1,6 @@
 use crate::board::{
-    get_empty_field_idx, get_shuffle_sequence, initialize_fields, touch_move, trigger_field,
-    PuzzleBoard,
+    get_empty_field_idx, get_shuffle_sequence, get_touch_direction, initialize_fields, touch_move,
+    trigger_field, PuzzleBoard,
 };
 use crate::expander::Expander;
 use crate::settings::SettingsBlock;
@@ -50,7 +50,7 @@ impl Component for SlidePuzzle {
         }
     }
 
-    fn update(&mut self, ctx: &Context<Self>, msg: Self::Message) -> bool {
+    fn update(&mut self, _ctx: &Context<Self>, msg: Self::Message) -> bool {
         log::info!("Received message {:?}", msg);
         match msg {
             SlidePuzzleMsg::Swap((a, b)) => match a != b {
@@ -95,7 +95,8 @@ impl Component for SlidePuzzle {
             },
             SlidePuzzleMsg::TouchStartCoords((x, y)) => match self.touch_start_coords {
                 Some(_) => {
-                    log::warn!("Cannot overwrite existing touchStart coordinates");
+                    log::warn!("Overwriting existing touchStart coordinates");
+                    self.touch_start_coords = Some((x, y));
                     false
                 }
                 None => {
@@ -275,46 +276,5 @@ impl SlidePuzzle {
             }
         });
         solve_callback
-    }
-}
-
-pub enum TouchMoveDirection {
-    Left,
-    Right,
-    Up,
-    Down,
-}
-
-fn get_touch_direction(
-    x_start: i32,
-    y_start: i32,
-    x_end: i32,
-    y_end: i32,
-) -> Option<TouchMoveDirection> {
-    let d_x = x_end - x_start;
-    let d_y = y_end - y_start;
-
-    if d_x.abs() + d_y.abs() < 40 {
-        // Overall displacement is too small, ignore
-        return None;
-    }
-
-    match d_x.abs() > d_y.abs() {
-        true => {
-            // Horizontal
-            if d_x > 0 {
-                return Some(TouchMoveDirection::Right);
-            } else {
-                return Some(TouchMoveDirection::Left);
-            }
-        }
-        false => {
-            // Vertical
-            if d_y > 0 {
-                return Some(TouchMoveDirection::Down);
-            } else {
-                return Some(TouchMoveDirection::Up);
-            }
-        }
     }
 }
